% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCBOV.R
\name{CCBOVin}
\alias{CCBOVin}
\title{EvaluaciÓn Y Determinación de la  Condición Corporal en Bovinos.}
\usage{
CCBOVin(PESO1, PDESV, SEXO, ALTURA, PESOF, GD, DAL, Fsex)
}
\arguments{
\item{PESO1}{(vector) variable que representa el peso inicial}

\item{PDESV}{(vector) nombre de la variable que represesta el peso desvastado.}

\item{SEXO}{(vector) variable idependiente que representa el sexo del animal.}

\item{ALTURA}{(vector) variable independiente que representa el frame y/o altura}

\item{PESOF}{(vector ) variable que representa el peso final de la engorda.}

\item{GD}{(vector) Variable independiente que representa la ganancia diaria
de peso.}

\item{DAL}{(vector) variable que hace referencia a los dias de engorda.}

\item{FSEXO}{(vector) variable idependiente que representa el Factor del sexo del animal,
es decir sies Hembra(0.8), Macho Entero (1.2) y Macho Castrado(1).}
}
\value{
Delvuelve una tabla con el promedio, desviacion estandar, minimo y
maximo de la Evaluacion de la Res, Tejido Adiposo(TA), Masa Muscular, Huesos,
Falacias, Tendones(HFT); tambien devuelve un grafico de kg predichos vs kg observados
 de Tejido Adiposo y Masa Muscular.
 @export

 @examples
 \dontrun(
 # DIRECTORIO DE TRABAJO
RUTA <- "C:/Users/yasuv/OneDrive/Documentos/PAQ  R/BD1.csv"

#--------------------------------------------------------------
#EJEMPLO 1
#DATOS <- read.csv(RUTA)
# CARGO DE LIBRERIA
library(CCOMBOY)
CCBOVin(PESO1=DATOS$PESO1,PDESV=DATOS$PDESV,SEXO=DATOS$SEXO,ALTURA=DATOS$ALTURA,
PESOF=DATOS$PESO.FIN, GD=DATOS$GD, DAL=DATOS$DAL, Fsex=DATOS$F.SEXO)
#--------------------------------------------------------------
 #EJEMPLO 2
 PESO1	<-c(250,	140,	115,	154,	193,	120,	153,	178,	117,	129,	187,	165,	209,	150,	277,	315,	354,	395)
SEXO<- c("Macho",	"Hembra", "Hembra","Hembra","Hembra","Hembra","Hembra","Hembra","Hembra","Hembra","Hembra","Hembra","Hembra","Hembra","Macho","Machoc","Machoc","Machoc")
Fsex <- c(1.2,.8,.8,.8,.8,.8,.8,.8,.8,.8,.8,.8,.8,.8,1.2,1,1,1)
PDESV	<-c(145,130,112,150,190,110,150,173,112,127,180,160,204,147,274,310,350,390)
ALTURA  <-c(6,6,6,6,6,6,5,5,5,5,4,9,9,9,9,9,7,7)
PESOF	<- c(550,560,600,580,597,509,530,590,542,529,603,587,537,568,1050,950,940,970)
GD	<- c(1.5,1.6,1.7,1.4,1.3,1.1,1,1.2,1.5,1.9,1.1,1.3,1.2,1.4,1.2,1.3,1.6,1.7)
DA	<- c(135,135,135,135,135,135,135,135,135,135,135,135,135,135,135,135,135,135)

CCBOVin(PESO1=PESO1, PDESV=PDESV, SEXO=SEXO, ALTURA=ALTURA, PESOF=PESOF, GD=GD, DAL=DA, Fsex=Fsex)
 )
}
\description{
Realiza la evaluación y determina la Composicion Corporal en Bovinos (TA Y Masa Muscular)
}
